// gen/ts/src/nama/api/v1/media_pb.d.ts
export declare type EpisodePosition = Message<"nama.api.v1.EpisodePosition"> & {
  /**
   * @generated from field: uint32 season_number = 1;
   */
  seasonNumber: number;

  /**
   * @generated from field: uint32 episode_number = 2;
   */
  episodeNumber: number;
};

/**
 * Describes the message nama.api.v1.EpisodePosition.
 * Use `create(EpisodePositionSchema)` to create a new message.
 */
export declare const EpisodePositionSchema: GenMessage<EpisodePosition>;

/**
 * @generated from message nama.api.v1.MediaUserState
 */
export declare type MediaUserState = Message<"nama.api.v1.MediaUserState"> & {
  /**
   * @generated from field: bool watched = 1;
   */
  watched: boolean;

  /**
   * @generated from field: google.protobuf.Duration position = 2;
   */
  position?: Duration | undefined;

  /**
   * @generated from field: google.protobuf.Duration duration = 3;
   */
  duration?: Duration | undefined;

  /**
   * @generated from field: google.protobuf.Timestamp updated_at = 4;
   */
  updatedAt?: Timestamp | undefined;
};

/**
 * Describes the message nama.api.v1.MediaUserState.
 * Use `create(MediaUserStateSchema)` to create a new message.
 */
export declare const MediaUserStateSchema: GenMessage<MediaUserState>;

/**
 * @generated from message nama.api.v1.VideoQuality
 */
export declare type VideoQuality = Message<"nama.api.v1.VideoQuality"> & {
  /**
   * @generated from field: string codec = 1;
   */
  codec: string;

  /**
   * @generated from field: optional uint32 width = 2;
   */
  width?: number | undefined;

  /**
   * @generated from field: optional uint32 height = 3;
   */
  height?: number | undefined;

  /**
   * @generated from field: optional nama.api.v1.DynamicRange dynamic_range = 4;
   */
  dynamicRange?: DynamicRange | undefined;
};

/**
 * Describes the message nama.api.v1.VideoQuality.
 * Use `create(VideoQualitySchema)` to create a new message.
 */
export declare const VideoQualitySchema: GenMessage<VideoQuality>;

/**
 * @generated from message nama.api.v1.AudioQuality
 */
export declare type AudioQuality = Message<"nama.api.v1.AudioQuality"> & {
  /**
   * @generated from field: string codec = 1;
   */
  codec: string;

  /**
   * @generated from field: optional uint32 channel_count = 2;
   */
  channelCount?: number | undefined;

  /**
   * @generated from field: optional nama.api.v1.SpatialAudioFormat spatial_format = 3;
   */
  spatialFormat?: SpatialAudioFormat | undefined;
};

/**
 * Describes the message nama.api.v1.AudioQuality.
 * Use `create(AudioQualitySchema)` to create a new message.
 */
export declare const AudioQualitySchema: GenMessage<AudioQuality>;

/**
 * @generated from message nama.api.v1.MediaSourceSummary
 */
export declare type MediaSourceSummary = Message<"nama.api.v1.MediaSourceSummary"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: optional string label = 2;
   */
  label?: string | undefined;

  /**
   * @generated from field: bool is_default = 3;
   */
  isDefault: boolean;

  /**
   * @generated from field: nama.api.v1.SourceAvailability availability = 4;
   */
  availability: SourceAvailability;

  /**
   * @generated from field: optional string container = 5;
   */
  container?: string | undefined;

  /**
   * @generated from field: nama.api.v1.VideoQuality video_quality = 6;
   */
  videoQuality?: VideoQuality | undefined;

  /**
   * @generated from field: nama.api.v1.AudioQuality audio_quality = 7;
   */
  audioQuality?: AudioQuality | undefined;
};

/**
 * Describes the message nama.api.v1.MediaSourceSummary.
 * Use `create(MediaSourceSummarySchema)` to create a new message.
 */
export declare const MediaSourceSummarySchema: GenMessage<MediaSourceSummary>;

/**
 * @generated from message nama.api.v1.ArtworkReference
 */
export declare type ArtworkReference = Message<"nama.api.v1.ArtworkReference"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: nama.api.v1.ArtworkRole role = 2;
   */
  role: ArtworkRole;

  /**
   * @generated from field: optional uint32 width = 3;
   */
  width?: number | undefined;

  /**
   * @generated from field: optional uint32 height = 4;
   */
  height?: number | undefined;

  /**
   * @generated from field: optional string locale = 5;
   */
  locale?: string | undefined;

  /**
   * @generated from field: nama.api.v1.ArtworkTextPresence text_presence = 6;
   */
  textPresence: ArtworkTextPresence;
};

/**
 * Describes the message nama.api.v1.ArtworkReference.
 * Use `create(ArtworkReferenceSchema)` to create a new message.
 */
export declare const ArtworkReferenceSchema: GenMessage<ArtworkReference>;

/**
 * @generated from message nama.api.v1.MediaSummary
 */
export declare type MediaSummary = Message<"nama.api.v1.MediaSummary"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: nama.api.v1.MediaKind kind = 2;
   */
  kind: MediaKind;

  /**
   * @generated from field: string title = 3;
   */
  title: string;

  /**
   * @generated from field: optional uint32 release_year = 4;
   */
  releaseYear?: number | undefined;

  /**
   * @generated from field: google.protobuf.Duration runtime = 5;
   */
  runtime?: Duration | undefined;

  /**
   * @generated from field: optional string content_rating = 6;
   */
  contentRating?: string | undefined;

  /**
   * @generated from field: optional string primary_genre = 7;
   */
  primaryGenre?: string | undefined;

  /**
   * @generated from field: nama.api.v1.EpisodePosition episode_position = 8;
   */
  episodePosition?: EpisodePosition | undefined;

  /**
   * @generated from field: repeated nama.api.v1.ArtworkReference artwork = 9;
   */
  artwork: ArtworkReference[];

  /**
   * @generated from field: nama.api.v1.MediaUserState user_state = 10;
   */
  userState?: MediaUserState | undefined;

  /**
   * @generated from field: nama.api.v1.Playability playability = 11;
   */
  playability: Playability;

  /**
   * @generated from field: nama.api.v1.MediaSourceSummary default_source = 12;
   */
  defaultSource?: MediaSourceSummary | undefined;
};
// ...
export enum MediaKind {
  /**
   * @generated from enum value: MEDIA_KIND_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: MEDIA_KIND_MOVIE = 1;
   */
  MOVIE = 1,

  /**
   * @generated from enum value: MEDIA_KIND_SHOW = 2;
   */
  SHOW = 2,

  /**
   * @generated from enum value: MEDIA_KIND_SEASON = 3;
   */
  SEASON = 3,

  /**
   * @generated from enum value: MEDIA_KIND_EPISODE = 4;
   */
  EPISODE = 4,
}

/**
 * Describes the enum nama.api.v1.MediaKind.
 */
export declare const MediaKindSchema: GenEnum<MediaKind>;

/**
 * @generated from enum nama.api.v1.Playability
 */
export enum Playability {
  /**
   * @generated from enum value: PLAYABILITY_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: PLAYABILITY_PLAYABLE = 1;
   */
  PLAYABLE = 1,

  /**
   * @generated from enum value: PLAYABILITY_TEMPORARILY_UNAVAILABLE = 2;
   */
  TEMPORARILY_UNAVAILABLE = 2,

  /**
   * @generated from enum value: PLAYABILITY_NO_AVAILABLE_SOURCE = 3;
   */
  NO_AVAILABLE_SOURCE = 3,
}

/**
 * Describes the enum nama.api.v1.Playability.
 */
export declare const PlayabilitySchema: GenEnum<Playability>;

/**
 * @generated from enum nama.api.v1.SourceAvailability
 */
export enum SourceAvailability {
  /**
   * @generated from enum value: SOURCE_AVAILABILITY_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: SOURCE_AVAILABILITY_AVAILABLE = 1;
   */
  AVAILABLE = 1,

  /**
   * @generated from enum value: SOURCE_AVAILABILITY_PROVIDER_UNAVAILABLE = 2;
   */
  PROVIDER_UNAVAILABLE = 2,

  /**
   * @generated from enum value: SOURCE_AVAILABILITY_UNSUPPORTED = 3;
   */
  UNSUPPORTED = 3,
}

/**
 * Describes the enum nama.api.v1.SourceAvailability.
 */
export declare const SourceAvailabilitySchema: GenEnum<SourceAvailability>;

/**
 * @generated from enum nama.api.v1.DynamicRange
 */
export enum DynamicRange {
  /**
   * @generated from enum value: DYNAMIC_RANGE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: DYNAMIC_RANGE_SDR = 1;
   */
  SDR = 1,

  /**
   * @generated from enum value: DYNAMIC_RANGE_HDR10 = 2;
   */
  HDR10 = 2,

  /**
   * @generated from enum value: DYNAMIC_RANGE_HDR10_PLUS = 3;
   */
  HDR10_PLUS = 3,

  /**
   * @generated from enum value: DYNAMIC_RANGE_HLG = 4;
   */
  HLG = 4,

  /**
   * @generated from enum value: DYNAMIC_RANGE_DOLBY_VISION = 5;
   */
  DOLBY_VISION = 5,
}

/**
 * Describes the enum nama.api.v1.DynamicRange.
 */
export declare const DynamicRangeSchema: GenEnum<DynamicRange>;

/**
 * @generated from enum nama.api.v1.SpatialAudioFormat
 */
export enum SpatialAudioFormat {
  /**
   * @generated from enum value: SPATIAL_AUDIO_FORMAT_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: SPATIAL_AUDIO_FORMAT_NONE = 1;
   */
  NONE = 1,

  /**
   * @generated from enum value: SPATIAL_AUDIO_FORMAT_DOLBY_ATMOS = 2;
   */
  DOLBY_ATMOS = 2,

  /**
   * @generated from enum value: SPATIAL_AUDIO_FORMAT_DTS_X = 3;
   */
  DTS_X = 3,
}
// ...
export enum ArtworkRole {
  /**
   * @generated from enum value: ARTWORK_ROLE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: ARTWORK_ROLE_POSTER = 1;
   */
  POSTER = 1,

  /**
   * @generated from enum value: ARTWORK_ROLE_BACKDROP = 2;
   */
  BACKDROP = 2,

  /**
   * @generated from enum value: ARTWORK_ROLE_LOGO = 3;
   */
  LOGO = 3,

  /**
   * @generated from enum value: ARTWORK_ROLE_THUMBNAIL = 4;
   */
  THUMBNAIL = 4,

  /**
   * @generated from enum value: ARTWORK_ROLE_PORTRAIT = 5;
   */
  PORTRAIT = 5,
}

/**
 * Describes the enum nama.api.v1.ArtworkRole.
 */
export declare const ArtworkRoleSchema: GenEnum<ArtworkRole>;

/**
 * @generated from enum nama.api.v1.ArtworkTextPresence
 */
export enum ArtworkTextPresence {
  /**
   * @generated from enum value: ARTWORK_TEXT_PRESENCE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: ARTWORK_TEXT_PRESENCE_UNKNOWN = 1;
   */
  UNKNOWN = 1,

  /**
   * @generated from enum value: ARTWORK_TEXT_PRESENCE_TEXTLESS = 2;
   */
  TEXTLESS = 2,

  /**
   * @generated from enum value: ARTWORK_TEXT_PRESENCE_CONTAINS_TEXT = 3;
   */
  CONTAINS_TEXT = 3,
}